[REF] Parser: move list parsing logic to List plugin
